add dailylimits modify

+++ src/AElf.CrossChainServer.Application.Contracts/CrossChain/CrossChainDailyLimitsDto.cs
@@ -3,5 +3,5 @@ namespace AElf.CrossChainServer.CrossChain;
 public class CrossChainDailyLimitsDto
 {
     public string Token { get; set; }
-    public long Allowance { get; set; }
+    public decimal Allowance { get; set; }
 }
